+++ CHANGELOG.md
@@ -20,7 +20,7 @@
     - Removed the distinction between "visible" sets and "selected" (i.e. checked) sets.
 - Upgrade deck.gl to 8.3 and viv to 0.5
 - Fix build issues with nebula.
-- Fix cell color encoding on GenesSubsrciber/Table when selecting cells.
+- Clear the cell set selection when selecting a gene in `GenesSubsrciber`. Clear the gene selection when selecting cell set(s) in `CellSetsManagerSubscriber`.
 
 ## [1.0.0](https://www.npmjs.com/package/vitessce/v/1.0.0) - 2020-09-2